+++ POS/POS.Data/Data/DeviceManagementRepository.cs
@@ -16,7 +16,7 @@ namespace POS.Infrastructure.Data
         }
         public async Task<MachineDto> GetConnectedMachine(string machineId, string ipAddress)
         {
-            const string sql = @"SELECT MACH_NO, CASINO_MACH_NO, MODEL_DESC, IP_ADDRESS, REMOVED_FLAG, BALANCE, LAST_ACTIVITY FROM [dbo].[MACH_SETUP] WHERE MACH_NO = @MachNo AND IP_ADDRESS = @IpAddress AND REMOVED_FLAG = @RemovedFlag";
+            const string sql = @"SELECT MACH_NO, CASINO_MACH_NO, MODEL_DESC, IP_ADDRESS, REMOVED_FLAG, BALANCE, LAST_ACTIVITY, ACTIVE_FLAG FROM [dbo].[MACH_SETUP] WHERE MACH_NO = @MachNo AND IP_ADDRESS = @IpAddress AND REMOVED_FLAG = @RemovedFlag";
             return await Db.SingleOrDefaultAsync<MachineDto>(sql,
                 new
                 {
@@ -28,7 +28,7 @@ namespace POS.Infrastructure.Data
 
         public async Task<IEnumerable<MachineDto>> GetConnectedMachines()
         {
-            const string sql = @"SELECT MACH_NO, CASINO_MACH_NO, MODEL_DESC, IP_ADDRESS, REMOVED_FLAG, BALANCE, LAST_ACTIVITY FROM [dbo].[MACH_SETUP] WHERE REMOVED_FLAG = @RemovedFlag";
+            const string sql = @"SELECT MACH_NO, CASINO_MACH_NO, MODEL_DESC, IP_ADDRESS, REMOVED_FLAG, BALANCE, LAST_ACTIVITY, ACTIVE_FLAG FROM [dbo].[MACH_SETUP] WHERE REMOVED_FLAG = @RemovedFlag";
             return await Db.FetchAsync<MachineDto>(sql, 
                 new
                  {